Log warning if br_netfilter cannot be enabled rather than fatally exiting
